requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
google-generativeai
python-multipart
supabase
python-jose       
passlib           
email-validator